refactor(cli): extract service settings and inference commands
Move service, settings, and inference command implementations out of run.rs while preserving compatibility re-exports.\n\nPart of #2304
